requirements.txt
discord.py>=2.4.0
yt-dlp>=2024.7.1
python-dotenv>=1.0.1
python-json-logger>=2.0.7
aiosqlite>=0.20.0
cryptography>=43.0.0
PyNaCl>=1.5.0
